Extract ProposalsList from InitiativeCards
