+++ front/app/components/InitiativeCards/index.tsx
@@ -8,7 +8,6 @@ import { trackEventByName } from 'utils/analytics';
 import tracks from './tracks';
 
 // components
-import InitiativeCard from 'components/InitiativeCard';
 import InitiativesMap from 'components/InitiativesMap';
 import { Spinner } from '@citizenlab/cl2-component-library';
 import SortFilterDropdown from './SortFilterDropdown';
@@ -52,8 +51,8 @@ import {
   defaultCardStyle,
 } from 'utils/styleUtils';
 import { ScreenReaderOnly } from 'utils/a11y';
-import { rgba } from 'polished';
 import EmptyProposals from './EmptyProposals';
+import ProposalsList from './ProposalsList';
 
 const gapWidth = 35;
 
@@ -143,51 +142,6 @@ const ContentLeft = styled.div`
   position: relative;
 `;
 
-const Loading = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  ${media.biggerThanMinTablet`
-    height: calc(100vh - 280px);
-    position: sticky;
-    top: 200px;
-  `}
-
-  ${media.smallerThanMinTablet`
-    height: 150px;
-  `}
-`;
-
-const InitiativesList = styled.div`
-  margin-left: -13px;
-  margin-right: -13px;
-  display: flex;
-  flex-wrap: wrap;
-`;
-
-const StyledInitiativeCard = styled(InitiativeCard)`
-  flex-grow: 0;
-  width: calc(100% * (1 / 3) - 26px);
-  margin-left: 13px;
-  margin-right: 13px;
-
-  @media (max-width: 1440px) and (min-width: 1279px) {
-    width: calc(100% * (1 / 3) - 16px);
-    margin-left: 8px;
-    margin-right: 8px;
-  }
-
-  @media (max-width: 1279px) and (min-width: 768px) {
-    width: calc(100% * (1 / 2) - 26px);
-  }
-
-  ${media.smallerThanMinTablet`
-    width: 100%;
-  `};
-`;
-
 const ContentRight = styled.div<{ filterColumnWidth: number }>`
   flex: 0 0 ${({ filterColumnWidth }) => filterColumnWidth}px;
   width: ${({ filterColumnWidth }) => filterColumnWidth}px;
@@ -251,21 +205,6 @@ const Spacer = styled.div`
 const StyledViewButtons = styled(ViewButtons)`
   margin-right: 20px;
 `;
-
-const Footer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  margin-top: 30px;
-
-  ${media.smallerThanMinTablet`
-    flex-direction: column;
-    align-items: stretch;
-    margin-top: 0px;
-  `}
-`;
-
-const ShowMoreButton = styled(Button)``;
 
 interface InputProps {
   className?: string;
@@ -470,10 +409,9 @@ class InitiativeCards extends PureComponent<Props & InjectedIntlProps, State> {
       initiativesFilterCounts,
       windowSize,
       className,
-      theme,
       invisibleTitleMessage,
     } = this.props;
-    const { list, hasMore, querying, loadingMore } = initiatives;
+    const { list, querying } = initiatives;
     const hasInitiatives = !isNilOrError(list) && list.length > 0;
     const biggerThanLargeTablet =
       windowSize && windowSize >= viewportWidths.largeTablet;
@@ -647,44 +585,7 @@ class InitiativeCards extends PureComponent<Props & InjectedIntlProps, State> {
 
             <Content>
               <ContentLeft>
-                {selectedView === 'card' &&
-                  !querying &&
-                  hasInitiatives &&
-                  list && (
-                    <InitiativesList id="e2e-initiatives-list">
-                      {list.map((initiative) => (
-                        <StyledInitiativeCard
-                          key={initiative.id}
-                          initiativeId={initiative.id}
-                        />
-                      ))}
-                    </InitiativesList>
-                  )}
-
-                {selectedView === 'card' && !querying && hasMore && (
-                  <Footer>
-                    <ShowMoreButton
-                      id="e2e-initiative-cards-show-more-button"
-                      onClick={this.loadMore}
-                      buttonStyle="secondary"
-                      text={<FormattedMessage {...messages.showMore} />}
-                      processing={loadingMore}
-                      height="50px"
-                      icon="showMore"
-                      iconPos="left"
-                      textColor={theme.colorText}
-                      bgColor={rgba(theme.colorText, 0.08)}
-                      bgHoverColor={rgba(theme.colorText, 0.12)}
-                      fontWeight="500"
-                    />
-                  </Footer>
-                )}
-
-                {selectedView === 'card' && querying && (
-                  <Loading id="initiatives-loading">
-                    <Spinner />
-                  </Loading>
-                )}
+                {selectedView === 'card' && <ProposalsList />}
 
                 {selectedView === 'map' && <InitiativesMap />}
